requirements.txt
matplotlib
ProDy==2.2.0
seaborn
stmol
streamlit-aggrid
streamlit-ext